+++ src/gras/Virtu/virtu_private.h
@@ -34,8 +34,8 @@ typedef struct {
    */
   xbt_dynar_t moddata;
 	
-	long int pid; /* pid of process, only for SG */
-	long int ppid; /* ppid of process, only for SG */
+  int pid; /* pid of process, only for SG */
+  int ppid; /* ppid of process, only for SG */
 } gras_procdata_t;
 
 gras_procdata_t *gras_procdata_get(void);
